+++ extensions/obsidian/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Obsidian Changelog
 
+## [Add Bookmark Support  | Version 1.12.0] - 2023-06-012
+- Update Starred notes to Bookmarks
+
 ## [Several Quality of Life Improvements | Version 1.11.0] - 2023-02-05
 
 - Add `Silent Mode` preference for the `Append to Daily Note Command` which prevents the command from opening the daily note